third_party/blink/web_tests/fast/dom/shadow/shadow-dom-mouse-event-adjust-offset-expected.txt

Tests to ensure that MouseEvent's offsetX and offsetY are adjusted in re-targeting.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS eventRecords['parent']['offsetY'] is 5
PASS eventRecords['shadow-host']['offsetY'] is 5 + 20
PASS eventRecords['top']['offsetY'] is 5 + 20
PASS eventRecords['parent']['offsetX'] is 80 / 2
PASS eventRecords['shadow-host']['offsetX'] is 80 / 2 + 100
PASS eventRecords['top']['offsetX'] is 80 / 2 + 100
PASS successfullyParsed is true

TEST COMPLETE
